ai-service/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.6
python-dotenv==1.0.1
pydantic==2.6.1
google-generativeai==0.3.2
aiofiles==23.2.1
python-magic==0.4.27
jq==1.6.0
pytest==7.4.4
pytest-asyncio==0.23.4
httpx==0.26.0
